+++ packages/client/src/client-details.ts
@@ -1,4 +1,6 @@
-import { Device, OS, Sdk } from './gen/video/sfu/models/models';
+import { ClientDetails, Device, OS, Sdk } from './gen/video/sfu/models/models';
+import { isReactNative } from './helpers/platforms';
+import { UAParser } from 'ua-parser-js';
 
 let sdkInfo: Sdk | undefined;
 let osInfo: OS | undefined;
@@ -26,4 +28,34 @@ export const setDeviceInfo = (info: Device) => {
 
 export const getDeviceInfo = () => {
   return deviceInfo;
+};
+
+export const getClientDetails = (): ClientDetails => {
+  if (isReactNative()) {
+    // Since RN doesn't support web, sharing browser info is not required
+    return {
+      sdk: getSdkInfo(),
+      os: getOSInfo(),
+      device: getDeviceInfo(),
+    };
+  }
+
+  const userAgent = new UAParser(navigator.userAgent);
+  const { browser, os, device, cpu } = userAgent.getResult();
+  return {
+    sdk: getSdkInfo(),
+    browser: {
+      name: browser.name || navigator.userAgent,
+      version: browser.version || '',
+    },
+    os: {
+      name: os.name || '',
+      version: os.version || '',
+      architecture: cpu.architecture || '',
+    },
+    device: {
+      name: `${device.vendor || ''} ${device.model || ''} ${device.type || ''}`,
+      version: '',
+    },
+  };
 };