+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,27 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.233](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.232...@stream-io/video-react-native-sdk-0.0.1-alpha.233) (2023-07-03)
+
+
+
+## [0.0.1-alpha.232](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.231...@stream-io/video-react-native-sdk-0.0.1-alpha.232) (2023-07-03)
+
+
+
+## [0.0.1-alpha.231](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.230...@stream-io/video-react-native-sdk-0.0.1-alpha.231) (2023-06-30)
+
+
+### Bug Fixes
+
+* **react-native:** google signin user id and refactor app context to store login info ([#727](https://github.com/GetStream/stream-video-js/issues/727)) ([90435e0](https://github.com/GetStream/stream-video-js/commit/90435e04494ca9ed954512de5b658a38e10dc1f4))
+
+
+
+## [0.0.1-alpha.230](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.229...@stream-io/video-react-native-sdk-0.0.1-alpha.230) (2023-06-30)
+
+
+
 ## [0.0.1-alpha.229](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.1-alpha.228...@stream-io/video-react-native-sdk-0.0.1-alpha.229) (2023-06-30)